Ability to build by branch name, with test.
This PR adds the ability to build from the command line by passing a branch name.  For the test, I renamed the multiple_versions/v2 folder to multiple_versions/v2#master in order to simulate a branch name.  Because we already have other tests for pulling github repositories, and because the url identifying and downloading is done further up in the stack, this folder "hack" should serve as a sufficient test so long as the version selected by branch name is added to the docConfig.versions in `make_final_doc_config.js`.