Simplified version to follow the configuring guide
This unifies the publishing and configuring guide, so the publishing guide is a continuation of what was setup in the configuring guide.

I simplified the sidebar layout, as well.  The contributing guide is now last in the list of guides.

Here's what's left to do before this can be merged:

 - [ ] Modify & Test the TravisCI shell scripts.  There are two scripts needed.  One for when there's a push to master.  Another for when there's a push to gh-pages.
 - [ ]  Move the folder setup process currently described here: https://github.com/bitovi/brain/wiki/ci.canjs.com into the shell script for the build step of Strider.
 - [ ] Move the server setup stuff somewhere else.  Maybe to my own blog.