Fix race condition in creating the static build
This fixes #243

The cause of the bug was that we were making the dist and build folders
at the same time that we were trying to copy the static.css into the
build folder. So the error would occur when things happened in this
order:

1. dist folder created
2. stat to check if dist/bundles/static.css exists (it won't)
3. copy the default static folder into the build folder
4. build folder is created

Error occurs when 3 occurs before 4.

Restructured the code so that we make the folders before trying the copy
procedure.