Fix error with demo tags on generated documentation
Why?
The previous release of documentjs upgrade to can-control@3.x but
it did not account for the fact that the DemoFrame control was
instantiated with a jquery element which is no longer natively supported
in 3.x; this cause can-event to try to call addEventListener on a
jquery object which threw an error.

This is fixed by:
Make sure the control is instatiated with an HtmlElement instead of
a jquery-wrapped element.

Closes #287